+++ source/ci_source/_tests/fixtures/dummy_ci.js
@@ -1,12 +1,21 @@
 class DummyCI {
-  get name() { return "Dummy Testing CI" }
+  get name() {
+    return "Dummy Testing CI"
+  }
 
-  get isCI() { return false }
-  get isPR() { return true }
+  get isCI() {
+    return false
+  }
+  get isPR() {
+    return true
+  }
 
-  get pullRequestID() { return this.env.pr }
-  get repoSlug() { return this.env.repo }
-  get supportedPlatforms() { return ["github"] }
+  get pullRequestID() {
+    return this.env.pr
+  }
+  get repoSlug() {
+    return this.env.repo
+  }
 }
 
 module.exports = DummyCI