+++ source/ci_source/providers/Buildkite.ts
@@ -51,10 +51,6 @@ export class Buildkite implements CISource {
     return this._parseRepoURL()
   }
 
-  get supportedPlatforms(): string[] {
-    return ["github"]
-  }
-
   get ciRunURL() {
     return process.env.BUILDKITE_BUILD_URL
   }