+++ source/platforms/git/diffToGitJSONDSL.ts
@@ -1,13 +1,14 @@
 import * as parseDiff from "parse-diff"
 import { includes } from "lodash"
 import { GitCommit } from "../../dsl/Commit"
+import { GitJSONDSL } from "../../dsl/GitDSL"
 
 /**
  * This function is essentially a "go from a diff to some simple structured data"
  * it's the steps needed for danger process.
  */
 
-export const diffToGitJSONDSL = (diff: string, commits: GitCommit[]) => {
+export const diffToGitJSONDSL = (diff: string, commits: GitCommit[]): GitJSONDSL => {
   const fileDiffs: any[] = parseDiff(diff)
 
   const addedDiffs = fileDiffs.filter((diff: any) => diff["new"])