+++ source/runner/templates/_tests/_bitbucketServerTemplate.test.ts
@@ -1,0 +1,40 @@
+import {
+  emptyResults,
+  failsResultsWithoutMessages,
+  warnResults,
+  failsResults,
+  summaryResults,
+} from "../../_tests/fixtures/ExampleDangerResults"
+import { template } from "../bitbucketServerTemplate"
+
+describe("generating messages for BitBucket server", () => {
+  it("shows no sections for empty results", () => {
+    const issues = template("blankID", emptyResults)
+    expect(issues).not.toContain("Fails")
+    expect(issues).not.toContain("Warnings")
+    expect(issues).not.toContain("Messages")
+  })
+
+  it("shows no sections for results without messages", () => {
+    const issues = template("blankID", failsResultsWithoutMessages)
+    expect(issues).not.toContain("Fails")
+    expect(issues).not.toContain("Warnings")
+    expect(issues).not.toContain("Messages")
+  })
+
+  it("Shows the failing messages in a section", () => {
+    const issues = template("blankID", failsResults)
+    expect(issues).toContain("Fails")
+    expect(issues).not.toContain("Warnings")
+  })
+
+  it("Shows the warning messages in a section", () => {
+    const issues = template("blankID", warnResults)
+    expect(issues).toContain("Warnings")
+    expect(issues).not.toContain("Fails")
+  })
+
+  it("summary result matches snapshot", () => {
+    expect(template("blankID", summaryResults)).toMatchSnapshot()
+  })
+})